Update submit behavior to populate list of movies
Created searchOmdb - a callback based function for getting search
results from the omdb API - it has an onLoad parameter for handling
the results.

The eventListener for the form uses the value of the search text box
to construct the url for hitting the search API. Once the results
are returned it calls the success callback with results.

The success callback currently creates an `li` element and populates
them into the movie list `ul` element.