BugFix: Update server.js to use post method
Previously, both favorites REST endpoints were GET - now the one used
to add new favorites is a POST, which is a better representation of
the functionality required of the endpoint. POST is used to add
resources to an already existing source (in this case, adding a favorite
to the existing database, stored in a file on the system).